[refactor] Move existing non W3C specification compliant XQuery Update implementation from org.exist.xquery.update to org.exist.xquery.update.legacy
