+++ exist-core/src/main/antlr/org/exist/xquery/parser/XQueryTree.g
@@ -53,6 +53,7 @@ header {
     import org.exist.xquery.value.*;
     import org.exist.xquery.functions.fn.*;
     import org.exist.xquery.update.*;
+    import org.exist.xquery.update.legacy.*;
     import org.exist.storage.ElementValue;
     import org.exist.xquery.functions.map.MapExpr;
     import org.exist.xquery.functions.array.ArrayConstructor;
