+++ exist-core/src/main/java/org/exist/xquery/update/legacy/Replace.java
@@ -19,7 +19,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.exist.xquery.update;
+package org.exist.xquery.update.legacy;
 
 import org.exist.EXistException;
 import org.exist.collections.triggers.TriggerException;
