+++ exist-core/src/test/java/org/exist/xquery/update/legacy/IndexIntegrationTest.java
@@ -19,7 +19,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.exist.xquery.update;
+package org.exist.xquery.update.legacy;
 
 import org.easymock.IArgumentMatcher;
 import org.easymock.IMocksControl;
