+++ exist-core/src/test/java/org/exist/xquery/update/legacy/UpdateValueTest.java
@@ -19,7 +19,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.exist.xquery.update;
+package org.exist.xquery.update.legacy;
 
 import org.junit.Test;
 import org.xmldb.api.base.XMLDBException;
